% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burstiness.R
\name{burstiness}
\alias{burstiness}
\title{burstiness()}
\usage{
burstiness(onsets)
}
\arguments{
\item{onsets}{List of onset time stamps from temporal data.}
}
\value{
A burstiness value B. Qualifier of bursty, periodic, or random.
}
\description{
burstiness()
}
\examples{
burstiness(c(0,1,5,8,10))
}
