% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusters.R
\name{create_clusters}
\alias{create_clusters}
\title{create_clusters()}
\usage{
create_clusters(onsets)
}
\arguments{
\item{onsets}{List of onset time stamps from temporal data.}
}
\value{
Histogram of IEIs. Cut-off point for determining which events get clustered.
}
\description{
create_clusters()
}
\examples{
create_clusters(c(0,1,5,7,9,24))
}
