% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_learninglab.R
\name{theme_learninglab}
\alias{theme_learninglab}
\title{theme_learninglab()}
\usage{
theme_learninglab()
}
\value{
Add onto a ggplot object for formatting.
}
\description{
theme_learninglab()
}
\examples{
ggplot(data, aes(x = independent)) +
geom_histogram() +
theme_learninglab()

}
